fix(tests): make all the tests pass (#34)
When cloning the repo, users expect to (or should expect to) be able to
run `go test ./...` and run all the tests. If running all the tests on a
fresh clone of main doesn't work, then it's hard to contribute. This
ensures that all of the tests run with `go test ./... -short`. The short
flag is used because there are some tests that are "costly" in terms of
time or environment. This is a well-known flag in Go for bypassing these
tests to provide a good onramp for new contributors.

Also, where one test was failing documentation has been improved and
failing tests were fixed.

---------

Signed-off-by: Anubhab Majumdar <anmajumdar@microsoft.com>
Signed-off-by: Evan Baker <rbtr@users.noreply.github.com>
Co-authored-by: Anubhab Majumdar <anmajumdar@microsoft.com>
Co-authored-by: Evan Baker <rbtr@users.noreply.github.com>