feat: run e2e in azure cluster.  (#109)
This PR enables the execution of end-to-end integration tests for Retina
within a Kubernetes cluster. While the current implementation is
designed for an Azure Kubernetes Service (AKS) cluster, it also supports
execution in other cloud environments.

Please note that these tests cannot be run on a local cluster such as
Kind due to its lack of support for packetparser eBPF programs.

The gh action needs to be run manually on pr before being approved. 

Running tests can be found here:

https://github.com/microsoft/retina/actions/workflows/integration.yaml

---------

Signed-off-by: Jacques I Massa <jac.massa0908@gmail.com>