feat: speed up cross-platform workflows
Speeds up the cross-platform workflows by running them on the
`ubuntu-latest` runner instead of the `windows-latest` runner. Go builds
on the Windows runner are extremely slow. Instead of natively building
on Windows, the cross-platform workflows will now run on the Linux
runner with GOOS/GOARCH environment variables set to cross-compile.

Signed-off-by: Evan Baker <rbtr@users.noreply.github.com>