feat(cd): sign released container images and Helm chart (#160)
# Description

Sign released container images with Cosign and Helm chart with `helm
package --sign`

## Related Issue

If this pull request is related to any issue, please mention it here.
Additionally, make sure that the issue is assigned to you before
submitting this pull request.

Closes #141 

## Checklist

- [x] I have read the [contributing
documantation](https://retina.sh/docs/contributing).
- [x] I signed and signed-off the commits (`git commit -S -s ...`)
- [x] I have correctly attributed the author(s) of the code.
- [ ] I have tested the changes locally.
- [ ] I have followed the project's style guidelines.
- [ ] I have updated the documentation, if necessary.
- [ ] I have added tests, if applicable.

## Screenshots (if applicable)

N/A

## Additional Notes

In order for testing this change, please create a GPG private key and a
Cosign private key and deposit as GH secrets for this repo

---

Please refer to the [CONTRIBUTING.md](../CONTRIBUTING.md) file for more
information on how to contribute to this project.

---------

Signed-off-by: hainenber <dotronghai96@gmail.com>
Signed-off-by: Đỗ Trọng Hải <41283691+hainenber@users.noreply.github.com>
Signed-off-by: Evan Baker <rbtr@users.noreply.github.com>
Co-authored-by: Evan Baker <rbtr@users.noreply.github.com>