fix: add secondary IPs for enrichment (#272)
# Description

Please provide a brief description of the changes made in this pull
request.

## Related Issue
Fixes #224 

## Checklist

- [x] I have read the [contributing
documentation](https://retina.sh/docs/contributing).
- [x] I signed and signed-off the commits (`git commit -S -s ...`). See
[this
documentation](https://docs.github.com/en/authentication/managing-commit-signature-verification/about-commit-signature-verification)
on signing commits.
- [x] I have correctly attributed the author(s) of the code.
- [x] I have tested the changes locally.
- [x] I have followed the project's style guidelines.
- [x] I have updated the documentation, if necessary.
- [x] I have added tests, if applicable.

## Screenshots (if applicable) or Testing Completed

![image](https://github.com/microsoft/retina/assets/22489650/9a38aae1-ecd9-49fe-bfa8-4f5898aadd1f)



## Additional Notes
Both the previous implementation and this PR don't clear stale
IP-endpoint bindings immediately from the ipToEpKey Map.
E.g. When endpoint A changes from IP address "1.1.1.1" to "2.2.2.2", the
stale binding between the IP address "1.1.1.1" and endpoint A is still
in the ipToEpKey map until the IP address "1.1.1.1" gets reassigned to
another pod.
However, this shouldn't be a problem. The DNS is supposed to provide
up-to-date IP addresses for the sender and receiver so that any flow
event will always contain the latest IP addresses.