feat: select interface based on default route instead of hardcoded eth0 for packetparser plugin (#215)
# Description

Instead of using the hardcoded `eth0` interface for `packetparser` it
now tries to get the outgoing interface by utilizing the equavalent of
`ip route show default`.
This should get the current active outgoing interface

## Related Issue

Fixes
https://github.com/microsoft/retina/issues/115#issuecomment-2031313354

## Checklist

- [x] I have read the [contributing
documentation](https://retina.sh/docs/contributing).
- [x] I signed and signed-off the commits (`git commit -S -s ...`)
- [x] I have correctly attributed the author(s) of the code.
- [x] I have tested the changes locally.
- [x] I have followed the project's style guidelines.
- [x] I have updated the documentation, if necessary.
- [x] I have added tests, if applicable.

## Screenshots (if applicable) or Testing Completed

Tested by running self build image on a cluster which failed before with
the error in the related issue, now the plugin starts correctly and
exports metrics.

## Additional Notes

I actually want to maybe remove the tests because the tests for this
feature are only testing the library and not the code really.

---

Please refer to the [CONTRIBUTING.md](../CONTRIBUTING.md) file for more
information on how to contribute to this project.

---------

Signed-off-by: Mikhail Epifanov <7824786+mikhail5555@users.noreply.github.com>
Co-authored-by: Mikhail Epifanov <Mikhail5555@users.noreply.github.com>