Use IsReply instead of Reply
The latter is a deprecated field because it's a regular boolean. The
protobuf well-known type is more of a tri-state value as I understand
it. This reduces the incidence of false-negatives, where a packet is
determined to not be a reply because this value is "false" (because of
the zero value).