Fix oversize lines
One instance of this is just inherently long... partly due to a linter
disablement. The other one is a long function signature, which is still
a problem. At least it's broken out onto multiple lines so it's a little
easier to deal with, but tbh I'm not crazy about doing that with
function / method signatures in Go.