Break dependency on listener
This package has some very loose dependency on golang.org/x/sys/unix, so
we cannot use it for cross-compatibility with windows. The only thing we
really need from this package is an interface definition. This
reimplements the interface that we need, and then uses reflection to
deal with a problematic enum type in that interface. It's ugly but it
breaks the dependency.