Copy Hubble peer service to internal
The Hubble Peer service's imports transitively include
`golang.org/x/sys/unix`. Fortunately, the dependencies are on interfaces,
so it's possible to break this dependency by providing our own
implementation, which we do. Unfortunately, Go compiles entire packages
when they are imported, so the mere mention of these imports causes
`golang.org/x/sys/unix` to be compiled (with failing results on
Windows). Since the entire peer service is an implementation of a gRPC
service, it's possible to simply copy the entire thing and replace the
problematic interface with an identical one, making no reference to the
`manager` package containing the infectious unix import.