Copy Hubble server package into internal
In order to re-enable the local server, the `server` package from
Cilium needed to be copied over. This is because it depends on
`serveroption` which transitively imports `golang.org/x/sys/unix`. We
have already ported `serveroption` without the dependency on
`golang.org/x/sys/unix`, so it's simply a matter of copying the `server`
package and swapping that dependency for the internal one.