fix: apiserver watcher refresh shouldn't crash plugin manager  (#576)
# Description

This PR modifies the APIServer watcher to improve error handling and
optimize performance:

1. **Error Handling**: Instead of returning a resolution error to the
caller, watcher now retries the host lookup, after 3 refresh, if we
still can't resolve the apiserver IP, we bail and return the error to
the watcher manager.
2. **Optimization**: Moved the hostname retrieval to the initialization
function. This change ensures that the hostname is obtained once during
initialization and not on every refresh, as the hostname does not
change.

## Related Issue

[#565](https://github.com/microsoft/retina/issues/565)

If this pull request is related to any issue, please mention it here.
Additionally, make sure that the issue is assigned to you before
submitting this pull request.

## Checklist

- [x] I have read the [contributing
documentation](https://retina.sh/docs/contributing).
- [x] I signed and signed-off the commits (`git commit -S -s ...`). See
[this
documentation](https://docs.github.com/en/authentication/managing-commit-signature-verification/about-commit-signature-verification)
on signing commits.
- [x] I have correctly attributed the author(s) of the code.
- [x] I have tested the changes locally.
- [x] I have followed the project's style guidelines.
- [x] I have updated the documentation, if necessary.
- [ ] I have added tests, if applicable.

## Screenshots (if applicable) or Testing Completed

Please add any relevant screenshots or GIFs to showcase the changes
made.

## Additional Notes

Add any additional notes or context about the pull request here.

---

Please refer to the [CONTRIBUTING.md](../CONTRIBUTING.md) file for more
information on how to contribute to this project.