Fix overwriting products of go generate
`go generate` modifies the tree to contain generated code. Previously,
only parts of the Retina tree that were needed to run `go generate` were
copied into the container's tree, to isolate that step from the rest of
the build process. This leverages the docker cache if nothing needs to
be regenerated. However, the subsequent COPY . . overwrote the
byproducts of this generation step, obviating its work.

The solution is a bit convoluted, but is necessary given how Dockerfiles
work. The results of generating plugins are tar'ed and dropped in the
root of the container filesystem. Once the rest of the project tree is
copied into the container, this tarball forcibly overwrites the
./pkg/plugin tree with the generated byproducts.

Nothing can modify the tree in-between these states, so this should be a
safe operation to perform.