fix: prevent potential misconfiguration of metricsInterval (#284) (#638)
# Description

I have added a new field `metricsIntervalDuration` of type
`time.Duration`. It will super cede the `metricsInterval` configuration
of the same type. The older key is not removed to keep backward
compatibility. The change is done to prevent potential misconfiguration
that can happen by assigning a duration to `metricsInterval` when it is
converted to seconds.

## Related Issue

Issue #284 

## Checklist

- [X] I have read the [contributing
documentation](https://retina.sh/docs/contributing).
- [X] I signed and signed-off the commits (`git commit -S -s ...`). See
[this
documentation](https://docs.github.com/en/authentication/managing-commit-signature-verification/about-commit-signature-verification)
on signing commits.
- [X] I have correctly attributed the author(s) of the code.
- [X] I have tested the changes locally.
- [X] I have followed the project's style guidelines.
- [X] I have updated the documentation, if necessary.
- [X] I have added tests, if applicable.

## Screenshots (if applicable) or Testing Completed

The logs showing that retina agent runs successffully after making the
changes.

![retina-after-chnage](https://github.com/user-attachments/assets/b4b1c643-6ebe-421f-9d32-f0f3d5dac8f6)

This shows the log line that `metricsInterval` should not be used.

![retina-using-metrics-interval](https://github.com/user-attachments/assets/9345ac4e-339e-4818-9571-324c38124abd)


## Additional Notes

I was not able to us the `zap` logger as it is probably not initialized
when the `GetConfig` is called or I might be missing how to use it so i
used the go `log` library. Do let me know if we need to change that.

---

Please refer to the [CONTRIBUTING.md](../CONTRIBUTING.md) file for more
information on how to contribute to this project.

Signed-off-by: Ritwik Ranjan <ritwikranjan99@gmail.com>