Add ability to overlay config to GetConfig
In order to make it possible for specific fields to be overridden by
user-controllable config files, the GetConfig function must be modified
to support an additional configuration file that can be filtered to
specific fields. This modifies GetConfig in a backwards-compatible way
to support a FilteredConfig type, specifiying additional overlays with
allowable fields from those config overlays.