fix: restart if pktmon server closes (#798)
# Description

the grpc windows client is pretty noisy with random grpc connection
errors
 

![image](https://github.com/user-attachments/assets/9ac3f1b1-40b0-481a-bf15-b668ae5176d9)

we currently log them and emit them, since (during test) they don't
impact the full functionality of the stream. However, this specific
error
 


![image](https://github.com/user-attachments/assets/df1cc421-4304-4674-866c-fbd63d3b774b)

is indicative that the pktmon-server.exe has died for some reason.
Currently, if the goroutine that kicked off the server exits, the main
thread will keep trying to dial the server. This change is to return in
the main thread if the goroutine exits, by writing to the error channel
 
in that case retina windows will restart and setup all of the pktmon
bits again

## Related Issue

If this pull request is related to any issue, please mention it here.
Additionally, make sure that the issue is assigned to you before
submitting this pull request.

## Checklist

- [ ] I have read the [contributing
documentation](https://retina.sh/docs/contributing).
- [ ] I signed and signed-off the commits (`git commit -S -s ...`). See
[this
documentation](https://docs.github.com/en/authentication/managing-commit-signature-verification/about-commit-signature-verification)
on signing commits.
- [ ] I have correctly attributed the author(s) of the code.
- [ ] I have tested the changes locally.
- [ ] I have followed the project's style guidelines.
- [ ] I have updated the documentation, if necessary.
- [ ] I have added tests, if applicable.

## Screenshots (if applicable) or Testing Completed

Please add any relevant screenshots or GIFs to showcase the changes
made.

## Additional Notes

Add any additional notes or context about the pull request here.

---

Please refer to the [CONTRIBUTING.md](../CONTRIBUTING.md) file for more
information on how to contribute to this project.