fix(tests): Truncate username to 8 characters in TestE2ERetina
The clusterName variable in the TestE2ERetina function is generated using the curuser.Username. However, if the username is too long, it can cause issues. This commit refactors the code to truncate the username to 8 characters before generating the clusterName. Additionally, it adds a log message to inform about the truncation.

Temporarily fixes #880

Signed-off-by: Ritwik Ranjan <ritwikranjan@microsoft.com>