fix: Revert "deps: bump google.golang.org/grpc from 1.66.2 to 1.67.0 (#765)" (#932)
This reverts commit 0dc47e06ceaeff18939f341e481b653d4d5ef60a.

It was found, via bisecting, that this upgrade of gRPC caused #921. When
investigating that issue, it was found (using `openssl`) that agent pods
did successfully establish TLS connections using the certificates
generated by certgen. However, some issue existed with gRPC itself. The
protobufs for the peer service have not changed since ~2022, so this
upgrade is a sensible implication for the bug.

Closes #921