feat: Print capture output name when running retina capture create (#894)
# Description

This PR adds functionality to the `retina capture create` CLI to print
out the expected file output name of the capture.

This is done by shifting the logic of the timestamp to when the CLI
command is ran (rather than done after the capture is finished by the
job). This means that the timestamp must be passed around for the jobs
to actually generate the same filename as was presented to the user.

Screenshot below is an example of a capture job and the expected file
name, compared with the actual file name generated on the node.

## Related Issue

https://github.com/microsoft/retina/issues/286

## Checklist

- [x] I have read the [contributing
documentation](https://retina.sh/docs/contributing).
- [x] I signed and signed-off the commits (`git commit -S -s ...`). See
[this
documentation](https://docs.github.com/en/authentication/managing-commit-signature-verification/about-commit-signature-verification)
on signing commits.
- [x] I have correctly attributed the author(s) of the code.
- [x] I have tested the changes locally.
- [x] I have followed the project's style guidelines.
- [x] I have updated the documentation, if necessary.
- [x] I have added tests, if applicable.

## Screenshots (if applicable) or Testing Completed


![image](https://github.com/user-attachments/assets/76990e8d-7250-43dc-8c33-a02703793972)

---

Please refer to the [CONTRIBUTING.md](../CONTRIBUTING.md) file for more
information on how to contribute to this project.

---------

Signed-off-by: Kamil Przepiorowski <kamil.prz@gmail.com>