Remove `go run` from all `go generate` commands
Including `go run` in the //go:generate pragmas is never really how `go
generate` was intended to be used. While it's a seemingly-useful hack to
pin versions of tools, it breaks cross-compilation.

Cross-compilation breaks because if the requested GOARCH differs from
the one `go generate` is being run on, `go run` receives the same GOARCH
and builds the tool for that platform instead. The tool will never exec
correctly because of the platform mismatch.

This removes `go run` from all `//go:generate` pragmas and adds the
appropriate installation of these tools to the Dockerfile. Developers
are expected to have these tools installed in the same way in their
development environment (in the same way as go, clang, etc.).