Add glibc to operator's final image
As a consequence of removing CGO_ENABLED=0, we now require glibc in the
final runtime environment of both retina-agent and retina-operator.
`retina-agent` had this already by consequence of the inclusion of clang
and other machinery necessary to compile ebpf. `retina-operator`,
however, did not. This adds the contents of /lib and /usr/lib to the
final image in order to include glibc into `retina-operator`'s runtime
environment.