Do platform detection at runtime in KernelVersion
There's no particular reason why this KernelVersion platform detection
has to be done at compile time. It complicates testing, as the function
is redefined by build tags, so the only way to avoid invoking commands
on the host is to have a throwaway func var to assign KernelVersion to.

It's completely reasonable to think that there's other information that
we might want to grab from the host eventually, so this really should
reside in a client of some kind. This patch adds such a client and makes
KernelVersion a method of it. This method uses runtime.GOOS to decide at
runtime which command will be invoked based on the platform found.

This opens the door for using an interface to mock this method more
cleanly in tests that desire to do so. Those tests have been updated.

Finally, the global `KernelVersion` func has been removed in favor of
instantiating one of these HostInfoClients instead (the surface area of
`KernelVersion` was very small, so the refactor was trivial).