fix: Makefile typo (#1181)
# Description

Tiny fix to change a typo from `pod` to `prod`.

## Related Issue

[1070](https://github.com/microsoft/retina/issues/1070) - originally the
issue was for an error in building the docs, but was unable to replicate
the issue and marking it as closed. During investigation found this
typo.

## Checklist

- [x] I have read the [contributing
documentation](https://retina.sh/docs/contributing).
- [x] I signed and signed-off the commits (`git commit -S -s ...`). See
[this
documentation](https://docs.github.com/en/authentication/managing-commit-signature-verification/about-commit-signature-verification)
on signing commits.
- [x] I have correctly attributed the author(s) of the code.
- [x] I have tested the changes locally.
- [x] I have followed the project's style guidelines.
- [x] I have updated the documentation, if necessary.
- [x] I have added tests, if applicable.

---

Please refer to the [CONTRIBUTING.md](../CONTRIBUTING.md) file for more
information on how to contribute to this project.

Signed-off-by: Kamil <kamil.prz@gmail.com>