chore(docs): Add docs about architecture and fix structure (#1229)
# Description

The main goal of this PR was to document the Retina architecture, namely
the data plane and the available control planes. While working on this,
I also made some smaller changes to improve our consistency.

What this PR does:

Smaller changes

- Change the favicon image being used to one with a better aspect ratio.
See below for screenshots of before/after.
- The current logo looks stretched in browser tabs and bookmarks due to
the aspect ratio and not enough negative space.
- Restructure the way our headings and pages are organised. (This meant
a number of files had to be renamed/moved, whose content was not
touched.)
- For example currently if you click on "Captures" you navigate to a
landing page/readme, but if you click on "Installation" it opens up a
drop down of its pages. This is inconsistent. This PR makes all of the
headings open up by default. See screenshots below for before/after.
- Added subheadings to contributing page to make it easier to read.

Architecture docs

- A new "Architecture" page is added to the newly introduced
"Introduction" heading. This discusses the Data Plane and the available
Control Planes.
- As part of this a number of diagrams are introduced. Both the `.png`
and the source for the diagram `.excalidraw` are included so that the
diagrams can be versioned.
- Update the "What is Retina?" page to be consistent with the landing
page of `retina.sh` in terms of the feature descriptions. Also added a
"What is Hubble?" subsection with a brief description.

## Related Issue

https://github.com/microsoft/retina/issues/1055

There is also another PR opened for Hubble installation -
https://github.com/microsoft/retina/pull/1223

## Checklist

- [x] I have read the [contributing
documentation](https://retina.sh/docs/contributing).
- [x] I signed and signed-off the commits (`git commit -S -s ...`). See
[this
documentation](https://docs.github.com/en/authentication/managing-commit-signature-verification/about-commit-signature-verification)
on signing commits.
- [x] I have correctly attributed the author(s) of the code.
- [x] I have tested the changes locally.
- [x] I have followed the project's style guidelines.
- [x] I have updated the documentation, if necessary.
- [x] I have added tests, if applicable.

## Screenshots (if applicable) or Testing Completed

I think its a bit redundant to put the Architecture Diagrams in here, as
the files themselves are included. So check those out directly. Or open
up a preview of the markdown page itself in GitHub.


Retina favicon image

![image](https://github.com/user-attachments/assets/d6f5176f-de7b-4692-a5bf-889fd5f4614a)

Example of new file structure - showcasing that the heading now opens
rather than having a landing page when clicked on.

![image](https://github.com/user-attachments/assets/1a5c2fe1-a3bb-4b38-8702-1630bb49b8e7)

Parity of feature highlight between `retina.sh` and "What is Retina?"
page.

![image](https://github.com/user-attachments/assets/1dadf714-32e8-4a08-8b0e-5cf386f551dd)

---

Please refer to the [CONTRIBUTING.md](../CONTRIBUTING.md) file for more
information on how to contribute to this project.

---------

Signed-off-by: Kamil <kamil.prz@gmail.com>
Signed-off-by: Ritwik Ranjan <ritwikranjan99@gmail.com>
Signed-off-by: Alex Castilio dos Santos <alexsantos@microsoft.com>
Co-authored-by: Ritwik Ranjan <ritwikranjan99@gmail.com>
Co-authored-by: Alex Castilio <alexsantos@microsoft.com>