deps: bump github.com/safchain/ethtool from v0.5.10 to v0.6.0 (#1594)
# Description

This pull request updates the version of the `ethtool` dependency in the
`go.mod` file to ensure compatibility with the latest fixes. With go
version upgrade from 1.23 to 1.24, certain behavior of go with relation
memory management got changed (see
https://github.com/golang/go/issues/73536) which effected out linuxutil
plugin leading to a jump in memory usage. With the help of pprof
profiling we identified the root cause in ethtool library (see
https://github.com/golang/go/issues/73536). The ethtool lib owners
implemented a change to address the issue which was released with tag
`v0.6.0`.

Dependency update:

*
[`go.mod`](diffhunk://#diff-33ef32bf6c23acb95f5902d7097b7a1d5128ca061167ec0716715b0b9eeaa5f6L300-R300):
Updated the `github.com/safchain/ethtool` dependency from version
`v0.5.10` to `v0.6.0`.

## Screenshots (if applicable) or Testing Completed

I ran retina with a workload that was utilizing following network at
around ~30% cpu:

![image](https://github.com/user-attachments/assets/e8c74fe5-3317-4872-9543-46bd8486296b)

Here is the memory profile while running different versions of retina,
the last one is this commit.

![image](https://github.com/user-attachments/assets/bfd399e0-1ded-44d6-95e6-8830c6181c05)

While we don't come back to original level, we do see an improvement in
memory usage of around ~10%