fix: e2e job skipped due to yaml syntax error (#1614)
Our E2e job is getting skipped in the merge queue, because of YAML
parsing errors.
We have invalid syntax in the if-statements checking for success of
prerequisite jobs that need to run for `manifests` and `E2E` to run.
Looks like this used to be ignored and those jobs just ran, while
recently Github started enforcing stricter YAML syntax validation which
is causing these jobs to get skipped!
It started about a week ago, example errors at the bottom of the page:
https://github.com/microsoft/retina/actions/runs/14842770818

![image](https://github.com/user-attachments/assets/1906d486-e581-4ed3-8c93-baa8b3f15e0d)