chore(docs): Capture CLI docs (#1639)
# Description

Update the Setup and Capture docs to be more concise and clear.

Restructured the Setup page.

Added a `Capture with CRD` page for consistency - a little bit of
overlap with the Concepts/CRDs/Capture.

Renamed headings under Prometheus and Grafana to add "Configuration".

- The setup page had "Next steps: Prometheus & Grafana" at the bottom,
even though those were NOT the next steps according to the order of the
docs, so removed that text to avoid confusion.

## Related Issue

NA

## Checklist

- [x] I have read the [contributing
documentation](https://retina.sh/docs/Contributing/overview).
- [x] I signed and signed-off the commits (`git commit -S -s ...`). See
[this
documentation](https://docs.github.com/en/authentication/managing-commit-signature-verification/about-commit-signature-verification)
on signing commits.
- [x] I have correctly attributed the author(s) of the code.
- [x] I have tested the changes locally.
- [x] I have followed the project's style guidelines.
- [x] I have updated the documentation, if necessary.
- [x] I have added tests, if applicable.

## Screenshots (if applicable) or Testing Completed

### Setup page


![{9A23BD5F-AF49-4FBE-90A1-A3822BC71B3F}](https://github.com/user-attachments/assets/fb775360-3561-4dcc-95b6-0db72e859fde)

### Capture Overview page


![image](https://github.com/user-attachments/assets/85ab22e9-acec-4c4c-9a77-6c3ac3a9f3ee)

### Capture CLI page


![image](https://github.com/user-attachments/assets/5313fc57-969a-4136-822b-38b4bcc06b8e)

### Capture CRD page


![image](https://github.com/user-attachments/assets/e7d7fbc0-c731-46a8-95fd-bd82224b6772)


---

Please refer to the [CONTRIBUTING.md](../CONTRIBUTING.md) file for more
information on how to contribute to this project.

---------

Signed-off-by: Kamil <kamil.prz@gmail.com>